moved TrajectoryPrediction into its own file
